fix: do not use unique_hlgroup if config.animation is off
If config.animation is off, we should just use the existing hlgroup.
unique_hlgroups is only needed to dynamically create color groups during
animation. We cannot just overwrite the existing hlgroup because undo
and redo operations sometimes are performed in a quick manner and all
the colors for each operation will be in a state of clashing with each
other.