fix(animation.slide): fixes end_col out of range error in certain cases (#217)
* fix(animation.slide): fixes end_col out of range error in certain cases

Error to reproduce:

1. set `slide` to `cursor_moved` animation
2. add `cursor_moved` as an autocmd for `CursorMoved` event
3. prepare a buffer that has more than 20+ lines of code, and then yank
any 2 or above lines
4. quickly do `gg<S-v>G` to select all and `p` to paste the previous
yanked lines (Need to be very quick, it does'n happen if you slowly
select all and paste in a slow pace)
5. gets `end_col` is out of range error

---

This error happens because the slide animation are running both when you
select from top to bottom (more than 10 lines) and then when pasting the
yanked lines (replace the whole document with the yanked lines and
causes another 10 lines jump). Now it will try to run animation on the
end of the buffer (before paste) and without waiting for the animation,
the replaced text removed the line from the buffer and hence the error
occures.

---

To solve this, I have added more protected calls just for the slide to
prevent weird issue, and make sure to check the current end of file line
number and abort when the line is gone

* check it in single line block instead