fix(animation.slide): `ns_id` is invalid when switching buffer very quickly
This change should fix the issue where when we have multiple different
buffers opened, and also using cursor moved with slide animation, when
switching the buffer quickly, it will throw an error of `ns_id` is
invalid.

In this case, we added protected call to `del_extmark` in
`animate_clear` and do not fallback to `fade` animation when it could
not create an `extmark` in `slide` animation.